created slot-service

+++ slot-service/src/main/java/com/stackroute/slotservice/service/CreateSlotServiceImp.java
@@ -1,11 +1,24 @@
 package com.stackroute.slotservice.service;
-
+import com.stackroute.slotservice.exception.SlotNotFoundException;
 import com.stackroute.slotservice.model.CreateSlot;
+import com.stackroute.slotservice.model.DbSequence;
 import com.stackroute.slotservice.repository.CreateSlotRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.mongodb.core.MongoOperations;
+import org.springframework.data.mongodb.core.query.Criteria;
+import org.springframework.data.mongodb.core.query.Query;
+import org.springframework.data.mongodb.core.query.Update;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
+
 import java.util.HashSet;
+import java.util.List;
+import java.util.Objects;
+import java.util.Optional;
+
+import static org.springframework.data.mongodb.core.FindAndModifyOptions.options;
 
 
 @Service
@@ -17,28 +30,65 @@ public class CreateSlotServiceImp implements CreateSlotService{
     CreateSlotRepository createSlotRepository;
 
 
+    @Autowired
+    private MongoOperations mongoOperations;
 
-    @Override
-    public HashSet<CreateSlot> getAllSlots() {
-        createSlotRepository.findAll().forEach(createSlot -> CreatedSlotList.add(createSlot));
-        return CreatedSlotList;
+    public int getSequenceNumber(String sequenceName)
+    {
+        Query query = new Query(Criteria.where("id").is(sequenceName));
+        Update update = new Update().inc("seq", 1);
+        DbSequence counter = mongoOperations.findAndModify(query, update, options().returnNew(true).upsert(true), DbSequence.class);
+        return !Objects.isNull(counter)?counter.getSeq() : 1;
     }
 
-    @Override
-    public CreateSlot getSlotById(String id) {
-        return createSlotRepository.findById(id).get();
-    }
+
 
     @Override
-    public void addCreateSlot(CreateSlot createSlot) {
-        createSlotRepository.save(createSlot);
+    public ResponseEntity<?> getAllSlots() {
+        List<CreateSlot> createslot = createSlotRepository.findAll();
+        if(createslot != null && createslot.size() > 0){
+            return new ResponseEntity<List<CreateSlot>>(createslot, HttpStatus.OK);
+        } else {
+            return new ResponseEntity<String>("No slots found", HttpStatus.OK);
+        }
+    }
 
-    }
+
     @Override
-    public Boolean deleteCreatedSlotById(String id) {
-        createSlotRepository.deleteById(id);
-        return true;
+    public ResponseEntity<CreateSlot> getSlotById(Long id) {
+        Optional<CreateSlot> createSlot = createSlotRepository.findById(id);
+        if(createSlot.isPresent()){
+            return new ResponseEntity<>(createSlot.get(), HttpStatus.OK);
+        } else {
+            throw new SlotNotFoundException("slot with id " + id + " is not found.");
+        }
     }
+
+
+
+    @Override
+    public ResponseEntity<String> addCreateSlot(CreateSlot createSlot) {
+        CreateSlot savedSlot = createSlotRepository.save(createSlot);
+        if(savedSlot != null ){
+            return new ResponseEntity<>("slot added successfully.", HttpStatus.CREATED);
+        } else {
+            return new ResponseEntity<>("Could not create slot.", HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+    }
+
+
+    @Override
+    public ResponseEntity<String> deleteCreatedSlotById(Long id) {
+        Optional<CreateSlot> createSlot = createSlotRepository.findById(id);
+        System.out.println(createSlot);
+        if(createSlot.isPresent()){
+            createSlotRepository.deleteById(id);
+            return new ResponseEntity<>("Slot with id " + id + " deleted successfull", HttpStatus.OK);
+        } else {
+            throw new SlotNotFoundException("Slot with id " + id + " is not found.");
+        }
+    }
+
 
     @Override
     public void deleteAllCreateSlots() {
